backend/requirements.txt
slack-sdk>=3.27.0
pydantic-ai[google]>=0.0.49
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
